refactor: support for go library use case
