+++ core/render.go
@@ -1,63 +1,35 @@
-package core
+package wharf
 
 import (
-	"errors"
-	"flag"
 	"fmt"
+	"io"
 	"os"
 	"path/filepath"
-	"strings"
 	"text/template"
 
 	"gopkg.in/yaml.v3"
 )
 
-var defaultTemplateFileName = "Dockerfile.template"
-
-func init() {
-	v, ok := os.LookupEnv("DOCKER_RENDER_DEFAULT_TEMPLATE_FILE_NAME")
-	if ok {
-		defaultTemplateFileName = v
-	}
-}
-
-// const defaultContextPath =
-
-func Run() error {
-	os.Args = os.Args[1:]
-	valuesFilePath := flag.String("values", "./docker-values.yaml", "The path for the values file to use")
-	outputFilePath := flag.String("output", "", "The path for the output file")
-	templateFileName := flag.String("file-name", defaultTemplateFileName, "The name of the template file")
-	flag.Parse()
-	if flag.NArg() != 1 {
-		return errors.New("the path for the template file is required\n")
-	}
-	contextPath := flag.Arg(0)
-	template, err := getTemplate(contextPath, *templateFileName)
+// Render renders the template file in the given contextPath using the values files from the given path
+// It writes the rendered Dockerfile to the io.Writer passed in parameters. It returns an error if something goes wrong
+func Render(contextPath, templateFileName, valuesFilePath string, output io.Writer) error {
+	template, err := getTemplate(contextPath, templateFileName)
 	if err != nil {
 		return err
 	}
-	values, err := readValues(contextPath, *valuesFilePath)
+	values, err := readValues(contextPath, valuesFilePath)
 	if err != nil {
 		return err
 	}
-	err = render(*outputFilePath, template, values)
+
+	err = render(template, values, output)
 	return nil
 }
 
 // render renders the template and values to the file with the given path.
 // If something goes wrong it returns an error
-func render(path string, tmpl *template.Template, values map[string]any) error {
-	var file *os.File = os.Stdout
-	var err error
-	if len(strings.TrimSpace(path)) > 0 {
-		file, err = os.OpenFile(path, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0666)
-		if err != nil {
-			return err
-		}
-		defer file.Close()
-	}
-	err = tmpl.Execute(file, values)
+func render(tmpl *template.Template, values map[string]any, file io.Writer) error {
+	err := tmpl.Execute(file, values)
 	if err != nil {
 		return err
 	}
